Fix Simulator.readProximitySensor
There was a problem in readProximitySensor: it was never receiving the
correct information from the sensor (status was always 0).
The documentation says this: "Recommended operation modes for this
function are simx_opmode_streaming (the first call) and
simx_opmode_buffer (the following calls)". However, I could not make
work by first calling simx_opmode_streaming and then simx_opmode_buffer.

It seems to work with simx_opmode_streaming, but we might need to check
that part again in the future.